Move shared modules into a mixin/ directory
The seven non-cop modules lived alongside the cops in
lib/rubocop/cop/style/rbs_inline/, which caused two problems:

- `rake new_cop` injected the new require into the middle of the mixin
  prologue. RuboCop's RequireFileInjector only treats requires from the
  same directory as one sorted group, so the deliberately
  dependency-ordered mixin requires were part of the cop comparison set.
  A generated cop was required before file_filter, so adding the usual
  `prepend FileFilter` raised NameError at load time.
- missing_class_annotation.rb sat next to missing_data_class_annotation.rb
  with no entry in config/default.yml, so it read as an unregistered cop.

Separating the directory fixes both. The namespace is unchanged: the
modules stay under RuboCop::Cop::Style::RbsInline, matching how rubocop
(cop/mixin/) and rubocop-rspec (cop/rspec/mixin/) organize theirs. Keeping
it flat also keeps the cop bodies uniform, since RangeHelp and the other
mixins rubocop itself ships are spelled without a namespace.

Specs mirror the new layout. RSpec/SpecFilePathFormat is excluded for the
mixin spec directory, since the path deliberately no longer matches the
constant.

rbs_inline_cops.rb carries a comment explaining why the mixin and cop
requires form two blocks, next to the requires it governs. CLAUDE.md
points at the mixin directory instead of listing the modules: the list
only duplicated `ls` and would go stale, the same reason the section
already defers to config/default.yml for cops.
