+++ lib/rubocop/cop/style/rbs_inline/mixin/ast_utils.rb
