+++ lib/rubocop/cop/style/rbs_inline/mixin/class_comment_alignment.rb
