+++ lib/rubocop/cop/style/rbs_inline/mixin/comment_parser.rb
