+++ lib/rubocop/cop/style/rbs_inline/mixin/data_struct_helper.rb
