+++ lib/rubocop/cop/style/rbs_inline/mixin/file_filter.rb
