+++ lib/rubocop/cop/style/rbs_inline/mixin/missing_class_annotation.rb
