+++ lib/rubocop/cop/style/rbs_inline/mixin/source_code_helper.rb
