+++ sig/rubocop/cop/style/rbs_inline/mixin/ast_utils.rbs
@@ -1,4 +1,4 @@
-# Generated from lib/rubocop/cop/style/rbs_inline/ast_utils.rb with RBS::Inline
+# Generated from lib/rubocop/cop/style/rbs_inline/mixin/ast_utils.rb with RBS::Inline
 
 module RuboCop
   module Cop
