+++ sig/rubocop/cop/style/rbs_inline/mixin/class_comment_alignment.rbs
@@ -1,4 +1,4 @@
-# Generated from lib/rubocop/cop/style/rbs_inline/class_comment_alignment.rb with RBS::Inline
+# Generated from lib/rubocop/cop/style/rbs_inline/mixin/class_comment_alignment.rb with RBS::Inline
 
 module RuboCop
   module Cop
