+++ sig/rubocop/cop/style/rbs_inline/mixin/comment_parser.rbs
@@ -1,4 +1,4 @@
-# Generated from lib/rubocop/cop/style/rbs_inline/comment_parser.rb with RBS::Inline
+# Generated from lib/rubocop/cop/style/rbs_inline/mixin/comment_parser.rb with RBS::Inline
 
 module RuboCop
   module Cop
