+++ sig/rubocop/cop/style/rbs_inline/mixin/data_struct_helper.rbs
@@ -1,4 +1,4 @@
-# Generated from lib/rubocop/cop/style/rbs_inline/data_struct_helper.rb with RBS::Inline
+# Generated from lib/rubocop/cop/style/rbs_inline/mixin/data_struct_helper.rb with RBS::Inline
 
 module RuboCop
   module Cop
