+++ sig/rubocop/cop/style/rbs_inline/mixin/file_filter.rbs
@@ -1,4 +1,4 @@
-# Generated from lib/rubocop/cop/style/rbs_inline/file_filter.rb with RBS::Inline
+# Generated from lib/rubocop/cop/style/rbs_inline/mixin/file_filter.rb with RBS::Inline
 
 module RuboCop
   module Cop
