+++ sig/rubocop/cop/style/rbs_inline/mixin/missing_class_annotation.rbs
@@ -1,4 +1,4 @@
-# Generated from lib/rubocop/cop/style/rbs_inline/missing_class_annotation.rb with RBS::Inline
+# Generated from lib/rubocop/cop/style/rbs_inline/mixin/missing_class_annotation.rb with RBS::Inline
 
 module RuboCop
   module Cop
