+++ sig/rubocop/cop/style/rbs_inline/mixin/source_code_helper.rbs
@@ -1,4 +1,4 @@
-# Generated from lib/rubocop/cop/style/rbs_inline/source_code_helper.rb with RBS::Inline
+# Generated from lib/rubocop/cop/style/rbs_inline/mixin/source_code_helper.rb with RBS::Inline
 
 module RuboCop
   module Cop
