+++ spec/rubocop/cop/style/rbs_inline/mixin/file_filter_spec.rb
